pyproject.toml
[tool.poetry]
name = "file-system-py"
version = "0.1.0"
description = ""
authors = ["Raju Komati <komatiraju032@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{ include = "py-storage", from = "src" }]

[tool.poetry.dependencies]
python = "^3.9"
pydantic = "^2.0.2"
boto3 = "^1.28.0"
pyyaml = "^6.0"
boto3-type-annotations-with-docs = "^0.3.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
